+++ src/features/settings/common/constants/defaultSettings.ts
@@ -22,8 +22,10 @@ export const defaultSettings: Settings = {
     colorHomeCountry: false,
     colorVisitedCountries: true,
     colorUpcomingVisits: false,
+    numAtlasColors: 4,
     palettes: {
       standard: DEFAULT_COLOR_PALETTES.standard,
+      atlas: DEFAULT_COLOR_PALETTES.atlas,
       yearly: DEFAULT_COLOR_PALETTES.yearly,
       cumulative: DEFAULT_COLOR_PALETTES.cumulative,
     },