+++ src/layouts/shells/Sidebar/Sidebar.tsx
@@ -60,7 +60,7 @@ export function Sidebar() {
               sidebarExpanded ? t("sidebar.collapse") : t("sidebar.expand")
             }
             className="flex h-10 w-10 ms-1 hover:bg-sidebar-btn-hover transition"
-            icon={<FaBars className="text-2xl" />}
+            icon={<FaBars className="text-text text-2xl" />}
             rounded
           />
           {sidebarExpanded && (